requirements_simple.txt
# Основные зависимости для FastAPI проекта
fastapi==0.115.12
uvicorn==0.33.0
python-multipart==0.0.20
defusedxml==0.7.1

# База данных
sqlalchemy==1.4.39
sqlalchemy-firebird==0.7.5
fdb==2.0.1

# Pydantic и валидация
pydantic==2.4.2
pydantic_core==2.10.1

# Дополнительные утилиты
python-dotenv==1.0.0
requests==2.31.0
Jinja2==3.0.2

# Для работы с XML
lxml==4.9.3
xmltodict==0.13.0

# Для работы с файлами
aiofiles==23.2.1

# Системные зависимости
greenlet==1.1.2
typing_extensions==4.8.0
exceptiongroup==1.2.2
anyio==4.5.2
starlette==0.44.0
h11==0.14.0
sniffio==1.3.1
click==8.0.3
colorama==0.4.4
packaging==23.2
certifi==2020.6.20
charset-normalizer==3.3.2
idna==2.10
urllib3==1.25.10
six==1.15.0
setuptools==47.1.0